Extract cell to a completely separate class, specific to the View Controller
